Add project support
This adds support for projects, and also adds various flows needed for
project support.

* Make VirtualFile conform to ModificationTracker, and introduce VirtualFileManager to track changes
* Introduce messaging API to propagate information between components, so each component can handle its own logic without interweaving dependencies
* Introduce PklProjectManager, and data classes to model pkl projects. This component manages PklProject state, and persists it to `.pkl-lsp/projects.json`.
* Change resolvers to use PklProject context to look for details
* Introduce new ActionableNotification notification request, used to display messages with calls to action in the client
* Add "Sync Project" and "Configure Pkl CLI" flows
* Add logic to detect if PklProject files have been modified and project needs to be re-synced
* Introduce lifecycle events for Component (initialize, dispose) and managed by Project
* Move field `package` from `PklModule` to `VirtualFile`, to be consistent with where `pklProjectDir` is

+++ src/main/kotlin/org/pkl/lsp/util/CachedValuesManager.kt
@@ -24,7 +24,7 @@ data class CachedValue<T>(val value: T, val dependencies: List<ModificationTrack
 }
 
 class CachedValuesManager(project: Project) : Component(project) {
-  private val cachedValues: MutableMap<String, Pair<List<Int>, CachedValue<*>>> =
+  private val cachedValues: MutableMap<String, Pair<List<Long>, CachedValue<*>>> =
     ConcurrentHashMap()
 
   /** Returns the currently cached value, or `null` if the cached value is out of date. */
@@ -41,7 +41,7 @@ class CachedValuesManager(project: Project) : Component(project) {
     cachedValues[key] = value.dependencies.map { it.getModificationCount() } to value
   }
 
-  private fun CachedValue<*>.isUpToDate(lastModificationCounts: List<Int>): Boolean {
+  private fun CachedValue<*>.isUpToDate(lastModificationCounts: List<Long>): Boolean {
     if (lastModificationCounts.size != dependencies.size) {
       throw IllegalArgumentException("Modification counts do not match dependency size")
     }
